requirements.txt
absl-py==2.2.2
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.34.24
botocore==1.34.162
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cssselect==1.2.0
filelock==3.18.0
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.6.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.71.0
gunicorn==21.2.0
h5py==3.13.0
huggingface-hub==0.31.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
keras==2.15.0
libclang==18.1.1
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
nested-lookup==0.2.25
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
parsel==1.8.1
playwright==1.42.0
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==11.0.1
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.5.3
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-cpu==2.15.0
tensorflow-estimator==2.15.0
tensorflow==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
w3lib==2.3.1
Werkzeug==3.0.1
wrapt==1.14.1
cryptography>=41.0.0

